+++ src/f_fee_tui/general_command.py
@@ -1,7 +1,6 @@
 from queue import Queue
 
 from textual import events
-from textual import on
 from textual.app import ComposeResult
 from textual.widgets import Button
 from textual.widgets import Static
@@ -19,6 +18,10 @@ BTN_END_OBSERVATION = """
 End the current observation.
 """
 
+BTN_START_OBSERVATION = """
+Start a new observation with the given description.
+"""
+
 
 class GeneralCommand(Static):
 
@@ -28,17 +31,10 @@ class GeneralCommand(Static):
 
     def compose(self) -> ComposeResult:
         yield Button("Set FPGA Defaults", id='btn-set-fpga-defaults', classes='command')
+        yield Button("Start observation", id="btn-start-observation", classes='command')
         yield Button("End observation", id="btn-end-observation", classes='command')
 
     def _on_mount(self, event: events.Mount) -> None:
         self.query_one("#btn-set-fpga-defaults").tooltip = BTN_SET_FPGA_DEFAULTS
+        self.query_one("#btn-start-observation").tooltip = BTN_START_OBSERVATION
         self.query_one("#btn-end-observation").tooltip = BTN_END_OBSERVATION
-
-    @on(Button.Pressed, "#btn-end-observation")
-    def command_end_observation(self):
-        self.end_observation()
-
-    def end_observation(self):
-
-        self.notify("Ending the current observation")
-        self._command_q.put_nowait(("CM_CS", "end_observation", [], {}))